requirements.txt
streamlit==1.52.2
pandas==2.3.3
plotly==6.5.1
python-dotenv==1.2.1
requests==2.32.5
